add full MIT license text
By including a copy of the exact MIT license text, we make the terms of
the em-socksify license very clear.

Also, in the specific case of em-socksify's MIT license, all derivative
copies of em-socksify are required to contain the disclaimer text within
the MIT license (as described in the sentence "The above copyright
notice and this permission notice shall be included in all copies or
substantial portions of the Software.") So in order to ease development
and distribution of em-socksify, it makes sense to simply include this
text in the upstream project.